refactor: 아이콘 디렉토리 업데이트
