+++ apps/admin/src/assets/icons/SideMenu/Exit.tsx
