+++ apps/admin/src/assets/icons/SideMenu/Puzzle.tsx
