+++ apps/admin/src/assets/icons/index.ts
@@ -1,3 +1,3 @@
-export { default as ChevronsLeft } from './ChevronsLeft';
-export { default as Exit } from './Exit';
-export { default as Puzzle } from './Puzzle';
+export { default as ChevronsLeft } from './SideMenu/ChevronsLeft';
+export { default as Exit } from './SideMenu/Exit';
+export { default as Puzzle } from './SideMenu/Puzzle';
